requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Deep Learning & YOLO
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Jupyter Notebook
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Streamlit Deployment
streamlit>=1.28.0
streamlit-webrtc>=0.45.0

# Data Processing
scikit-image>=0.21.0
albumentations>=1.3.0

# Visualization
plotly>=5.17.0
tensorboard>=2.14.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Kaggle API (for dataset download)
kaggle>=1.5.0

# Logging & Monitoring
wandb>=0.15.0
loguru>=0.7.0
